curvenumber: an implementation of the USDA Soil Conservation Service Curve Number method

The SCS Curve Number (CN) is a well established method for the estimation of direct runoff from storm rainfall, developed by the USDA Soil Conservation Service and used in hydrologic engineering and environmental impact analyses. The curvenumber is an R package which allows to calculate: a) the direct storm runoff (Q), given rainfall (R) and CN number; b) the CN number, given rainfall (R) and runoff (Q) data; c) HOST-based soil classes mapping onto the CN soil classes. The package contains example data as well as a number of examples to test the main functionalities. The function EventIdentification(), for instance, is used to identify rainfall-runoff events and calculating the matching return period, according to Hjelmfelt (1980). The function CalculateCN() is used to identify Curve Number and k coefficient as well as to plot the CN-P asymptotic behaviour (see figure below), according to Hawkins (1993). The direct storm runoff can be calulated using the function CalculateQ() and CNfromMaps() implements the methodology illustrated in Bulygina et al. (2011), allowing to calculate the CN given at least the soil map of the area.

References

Bulygina, N., N. McIntyre, and H. Wheater (2011), Bayesian conditioning of a rainfall-runoff model for predicting flows in ungauged catchments and under land use changes, Water Resour. Res., 47, W02503, doi:10.1029/2010WR009240.

Hawkins, Richard H. 1993. Asymptotic Determination of Runoff Curve Numbers from Data. Journal of Irrigation and Drainage Engineering 119 (2). American Society of Civil Engineers: 334–45.

Hjelmfelt, Allen T. 1980. Empirical Investigation of Curve Number Technique. Journal of the Hydraulics Division 106 (9). ASCE: 1471–6.